Refactoring: added model/viewmodel for deleting components
